Move new project to FileMenuController
